chore: move devtools to bottom right (#160)
### TL;DR

Changed the position of TanStack Devtools from bottom-left to bottom-right.

### What changed?

Updated the TanStack Devtools configuration to display in the bottom-right corner of the screen instead of the bottom-left corner. The `position` property in the config object was modified from 'bottom-left' to 'bottom-right'.

### How to test?

1. Run the web application in development mode
2. Verify that the TanStack Devtools panel appears in the bottom-right corner of the screen
3. Confirm that the devtools functionality works as expected in the new position

### Why make this change?

The bottom-right position likely provides better visibility and reduces interference with other UI elements. This change improves the developer experience by positioning the devtools in a more convenient location.